Moved SimulationTest class to testutils package
