+++ src/test/java/uk/co/ramp/covid/simulation/population/InfantTest.java
@@ -4,7 +4,7 @@ import com.google.gson.JsonParseException;
 import org.junit.Test;
 import uk.co.ramp.covid.simulation.DailyStats;
 import uk.co.ramp.covid.simulation.testutil.PopulationGenerator;
-import uk.co.ramp.covid.simulation.util.SimulationTest;
+import uk.co.ramp.covid.simulation.testutil.SimulationTest;
 
 import java.util.List;
 import static org.junit.Assert.assertEquals;
