+++ src/test/java/uk/co/ramp/covid/simulation/population/PensionerTest.java
@@ -3,7 +3,7 @@ package uk.co.ramp.covid.simulation.population;
 import org.junit.Test;
 import uk.co.ramp.covid.simulation.DailyStats;
 import uk.co.ramp.covid.simulation.testutil.PopulationGenerator;
-import uk.co.ramp.covid.simulation.util.SimulationTest;
+import uk.co.ramp.covid.simulation.testutil.SimulationTest;
 
 import java.util.List;
